+++ src/components/common/MarkdownViewer.tsx
@@ -2,47 +2,14 @@
 
 import { useMemo } from "react";
 
+import { transformImagesSafely } from "@/utils/markdownImageUtils";
+
 import "@/styles/markdownViewer.css";
 import MarkdownPreview from "@uiw/react-markdown-preview";
 
 interface Props {
   content: string;
 }
-
-const sanitizeUrl = (raw: string) => {
-  let s = (raw || "").trim();
-
-  if (s.startsWith("<") && s.endsWith(">")) s = s.slice(1, -1);
-
-  try {
-    s = decodeURI(s);
-  } catch (e) {
-    console.warn("Invalid URI, skipping decode:", e);
-  }
-  s = encodeURI(s);
-  s = s.replace(
-    /[()'"<>]/g,
-    ch => "%" + ch.charCodeAt(0).toString(16).toUpperCase()
-  );
-  return s;
-};
-
-const transformImagesSafely = (md: string) => {
-  const parts = (md ?? "").split(/(```[\s\S]*?```)/g);
-  return parts
-    .map(part => {
-      if (part.startsWith("```")) return part;
-
-      return part.replace(
-        /!\[([^\]]*)\]\(\s*(?:<([^>]+)>|([^)\s]+))(?:\s+(?:"[^"]*"|'[^']*'|\([^)]*\)))?\s*\)/g,
-        (_, alt: string, urlA?: string, urlB?: string) => {
-          const url = sanitizeUrl(urlA || urlB || "");
-          return `![${alt}](${url})`;
-        }
-      );
-    })
-    .join("");
-};
 
 const MarkdownViewer = ({ content }: Props) => {
   const fixedContent = useMemo(
